Editor: Unify Header component
