+++ packages/e2e-test-utils/src/ensure-sidebar-opened.js
@@ -5,10 +5,9 @@
  */
 export async function ensureSidebarOpened() {
 	const toggleSidebarButton = await page.$(
-		'.edit-post-header__settings [aria-label="Settings"][aria-expanded="false"],' +
+		'.editor-header__settings [aria-label="Settings"][aria-expanded="false"],' +
 			'.edit-site-header__actions [aria-label="Settings"][aria-expanded="false"],' +
-			'.edit-widgets-header__actions [aria-label="Settings"][aria-expanded="false"],' +
-			'.edit-site-header-edit-mode__actions [aria-label="Settings"][aria-expanded="false"]'
+			'.edit-widgets-header__actions [aria-label="Settings"][aria-expanded="false"]'
 	);
 
 	if ( toggleSidebarButton ) {
