+++ packages/edit-post/src/components/header/index.js
@@ -1,24 +1,9 @@
 /**
- * External dependencies
- */
-import clsx from 'clsx';
-
-/**
  * WordPress dependencies
  */
-import {
-	DocumentBar,
-	PostSavedState,
-	PostPreviewButton,
-	store as editorStore,
-	privateApis as editorPrivateApis,
-} from '@wordpress/editor';
+import { privateApis as editorPrivateApis } from '@wordpress/editor';
 import { useSelect } from '@wordpress/data';
-import { useViewportMatch } from '@wordpress/compose';
 import { __unstableMotion as motion } from '@wordpress/components';
-import { store as preferencesStore } from '@wordpress/preferences';
-import { useState } from '@wordpress/element';
-import { store as blockEditorStore } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -29,21 +14,7 @@ import MainDashboardButton from './main-dashboard-button';
 import { store as editPostStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 
-const {
-	CollapsableBlockToolbar,
-	DocumentTools,
-	PostViewLink,
-	PreviewDropdown,
-	PinnedItems,
-	MoreMenu,
-	PostPublishButtonOrToggle,
-} = unlock( editorPrivateApis );
-
-const slideY = {
-	hidden: { y: '-50px' },
-	distractionFreeInactive: { y: 0 },
-	hover: { y: 0, transition: { type: 'tween', delay: 0.2 } },
-};
+const { Header: EditorHeader } = unlock( editorPrivateApis );
 
 const slideX = {
 	hidden: { x: '-100%' },
@@ -52,42 +23,17 @@ const slideX = {
 };
 
 function Header( { setEntitiesSavedStatesCallback, initialPost } ) {
-	const isWideViewport = useViewportMatch( 'large' );
-	const isLargeViewport = useViewportMatch( 'medium' );
-	const {
-		isTextEditor,
-		hasActiveMetaboxes,
-		isPublishSidebarOpened,
-		showIconLabels,
-		hasHistory,
-		hasFixedToolbar,
-		isZoomedOutView,
-	} = useSelect( ( select ) => {
-		const { get: getPreference } = select( preferencesStore );
-		const { getEditorMode } = select( editorStore );
-		const { __unstableGetEditorMode } = select( blockEditorStore );
-
+	const { hasActiveMetaboxes } = useSelect( ( select ) => {
 		return {
-			isTextEditor: getEditorMode() === 'text',
 			hasActiveMetaboxes: select( editPostStore ).hasMetaBoxes(),
-			hasHistory:
-				!! select( editorStore ).getEditorSettings()
-					.onNavigateToPreviousEntityRecord,
-			isPublishSidebarOpened:
-				select( editorStore ).isPublishSidebarOpened(),
-			showIconLabels: getPreference( 'core', 'showIconLabels' ),
-			hasFixedToolbar: getPreference( 'core', 'fixedToolbar' ),
-			isZoomedOutView: __unstableGetEditorMode() === 'zoom-out',
 		};
 	}, [] );
 
-	const hasTopToolbar = isLargeViewport && hasFixedToolbar;
-
-	const [ isBlockToolsCollapsed, setIsBlockToolsCollapsed ] =
-		useState( true );
-
 	return (
-		<div className="edit-post-header">
+		<EditorHeader
+			forceIsDirty={ hasActiveMetaboxes }
+			setEntitiesSavedStatesCallback={ setEntitiesSavedStatesCallback }
+		>
 			<MainDashboardButton.Slot>
 				<motion.div
 					variants={ slideX }
@@ -99,64 +45,8 @@ function Header( { setEntitiesSavedStatesCallback, initialPost } ) {
 					/>
 				</motion.div>
 			</MainDashboardButton.Slot>
-			<motion.div
-				variants={ slideY }
-				transition={ { type: 'tween', delay: 0.8 } }
-				className="edit-post-header__toolbar"
-			>
-				<DocumentTools disableBlockTools={ isTextEditor } />
-				{ hasTopToolbar && (
-					<CollapsableBlockToolbar
-						isCollapsed={ isBlockToolsCollapsed }
-						onToggle={ setIsBlockToolsCollapsed }
-					/>
-				) }
-				<div
-					className={ clsx( 'edit-post-header__center', {
-						'is-collapsed':
-							hasHistory &&
-							! isBlockToolsCollapsed &&
-							hasTopToolbar,
-					} ) }
-				>
-					{ hasHistory && <DocumentBar /> }
-				</div>
-			</motion.div>
-			<motion.div
-				variants={ slideY }
-				transition={ { type: 'tween', delay: 0.8 } }
-				className="edit-post-header__settings"
-			>
-				{ ! isPublishSidebarOpened && (
-					// This button isn't completely hidden by the publish sidebar.
-					// We can't hide the whole toolbar when the publish sidebar is open because
-					// we want to prevent mounting/unmounting the PostPublishButtonOrToggle DOM node.
-					// We track that DOM node to return focus to the PostPublishButtonOrToggle
-					// when the publish sidebar has been closed.
-					<PostSavedState forceIsDirty={ hasActiveMetaboxes } />
-				) }
-				<PreviewDropdown
-					disabled={ isZoomedOutView }
-					forceIsAutosaveable={ hasActiveMetaboxes }
-				/>
-				<PostPreviewButton
-					className="edit-post-header__post-preview-button"
-					forceIsAutosaveable={ hasActiveMetaboxes }
-				/>
-				<PostViewLink />
-				<PostPublishButtonOrToggle
-					forceIsDirty={ hasActiveMetaboxes }
-					setEntitiesSavedStatesCallback={
-						setEntitiesSavedStatesCallback
-					}
-				/>
-				{ ( isWideViewport || ! showIconLabels ) && (
-					<PinnedItems.Slot scope="core" />
-				) }
-				<MoreMenu />
-				<PostEditorMoreMenu />
-			</motion.div>
-		</div>
+			<PostEditorMoreMenu />
+		</EditorHeader>
 	);
 }
