+++ packages/edit-post/src/components/header/style.scss
@@ -1,251 +1,16 @@
-.edit-post-header {
-	height: $header-height;
-	background: $white;
-	display: flex;
-	flex-wrap: wrap;
-	align-items: center;
-	// The header should never be wider than the viewport, or buttons might be hidden. Especially relevant at high zoom levels. Related to https://core.trac.wordpress.org/ticket/47603#ticket.
-	max-width: 100vw;
-	justify-content: space-between;
-
-	// Make toolbar sticky on larger breakpoints
-	@include break-zoomed-in {
-		flex-wrap: nowrap;
-	}
-}
-
-.edit-post-header__toolbar {
-	display: flex;
-	// Allow this area to shrink to fit the toolbar buttons.
-	flex-shrink: 8;
-	// Take up the space of the toolbar so it can be justified to the left side of the toolbar.
-	flex-grow: 3;
-	// Hide the overflow so flex will limit its width. Block toolbar will allow scrolling on fixed toolbar.
-	overflow: hidden;
-	// Leave enough room for the focus ring to show.
-	padding: 2px 0;
-	align-items: center;
-	// Allow focus ring to be fully visible on furthest right button.
-	@include break-medium() {
-		padding-right: var(--wp-admin-border-width-focus);
-	}
-
-	.table-of-contents {
-		display: none;
-
-		@include break-small() {
-			display: block;
-		}
-	}
-}
-
-.edit-post-header__center {
-	flex-grow: 1;
-	display: flex;
-	justify-content: center;
-
-	&.is-collapsed {
-		display: none;
-	}
-}
-
 /**
- * Buttons on the right side
+ * Fullscreen mode button.
  */
-
-.edit-post-header__settings {
-	display: inline-flex;
-	align-items: center;
-	flex-wrap: nowrap;
-	padding-right: $grid-unit-05;
-
-	@include break-small () {
-		padding-right: $grid-unit-10;
-	}
-
-	gap: $grid-unit-10;
-}
-
-/**
- * Show icon labels.
- */
-
-.show-icon-labels.interface-pinned-items,
-.show-icon-labels .edit-post-header,
-.edit-post-header__dropdown {
-	.components-button.has-icon {
-		width: auto;
-
-		// Hide the button icons when labels are set to display...
-		svg {
-			display: none;
-		}
-		// ... and display labels.
-		&::after {
-			content: attr(aria-label);
-		}
-		&[aria-disabled="true"] {
-			background-color: transparent;
-		}
-	}
-	.is-tertiary {
-		&:active {
-			box-shadow: 0 0 0 1.5px var(--wp-admin-theme-color);
-			background-color: transparent;
-		}
-	}
-	// Exception for drodpdown toggle buttons.
+.show-icon-labels .editor-header {
 	// Exception for the fullscreen mode button.
-	.edit-post-fullscreen-mode-close.has-icon,
-	.components-button.has-icon.button-toggle {
-		svg {
-			display: block;
-		}
-		&::after {
-			content: none;
-		}
-	}
 	// Undo the width override for fullscreen mode button.
 	.edit-post-fullscreen-mode-close.has-icon {
 		width: $header-height;
-	}
-	// Don't hide MenuItemsChoice check icons
-	.components-menu-items-choice .components-menu-items__item-icon.components-menu-items__item-icon {
-		display: block;
-	}
-	.editor-document-tools__inserter-toggle.editor-document-tools__inserter-toggle,
-	.interface-pinned-items .components-button {
-		padding-left: $grid-unit;
-		padding-right: $grid-unit;
-
-		@include break-small {
-			padding-left: $grid-unit-15;
-			padding-right: $grid-unit-15;
+		svg {
+			display: block;
 		}
-	}
-
-	.editor-post-save-draft.editor-post-save-draft,
-	.editor-post-saved-state.editor-post-saved-state {
 		&::after {
 			content: none;
 		}
 	}
-}
-
-.show-icon-labels {
-	.edit-post-header__toolbar .block-editor-block-mover {
-		// Modified group borders.
-		border-left: none;
-
-		&::before {
-			content: "";
-			width: $border-width;
-			height: $grid-unit-30;
-			background-color: $gray-300;
-			margin-top: $grid-unit-05;
-			margin-left: $grid-unit;
-		}
-
-		// Modified block movers horizontal separator.
-		.block-editor-block-mover__move-button-container {
-			&::before {
-				width: calc(100% - #{$grid-unit-30});
-				background: $gray-300;
-				left: calc(50% + 1px);
-			}
-		}
-	}
-}
-
-.edit-post-header__dropdown {
-	.components-menu-item__button.components-menu-item__button,
-	.components-button.editor-history__undo,
-	.components-button.editor-history__redo,
-	.table-of-contents .components-button,
-	.components-button.block-editor-list-view {
-		margin: 0;
-		padding: 6px 6px 6px $grid-unit-50;
-		width: 14.625rem;
-		text-align: left;
-		justify-content: flex-start;
-	}
-}
-
-.show-icon-labels.interface-pinned-items {
-	padding: 6px $grid-unit-15 $grid-unit-15;
-	margin-top: 0;
-	margin-bottom: 0;
-	margin-left: -$grid-unit-15;
-	margin-right: -$grid-unit-15;
-	border-bottom: 1px solid $gray-400;
-	display: block;
-
-	> .components-button.has-icon {
-		margin: 0;
-		padding: 6px 6px 6px $grid-unit;
-		width: 14.625rem;
-		justify-content: flex-start;
-
-		&[aria-expanded="true"] svg {
-			display: block;
-			max-width: $grid-unit-30;
-		}
-		&[aria-expanded="false"] {
-			padding-left: $grid-unit-50;
-		}
-		svg {
-			margin-right: 8px;
-		}
-	}
-}
-
-.edit-post-header__post-preview-button {
-	@include break-small {
-		display: none;
-	}
-}
-
-.is-distraction-free {
-	.interface-interface-skeleton__header {
-		border-bottom: none;
-	}
-
-	.edit-post-header {
-		background-color: $white;
-		border-bottom: 1px solid #e0e0e0;
-		position: absolute;
-		width: 100%;
-
-
-		// hide some parts
-		& > .edit-post-header__settings > .edit-post-header__post-preview-button {
-			visibility: hidden;
-		}
-
-		& > .edit-post-header__toolbar .editor-document-tools__document-overview-toggle,
-		& > .edit-post-header__settings > .editor-preview-dropdown,
-		& > .edit-post-header__settings > .interface-pinned-items {
-			display: none;
-		}
-
-	}
-
-	// We need ! important because we override inline styles
-	// set by the motion component.
-	.interface-interface-skeleton__header:focus-within {
-		opacity: 1 !important;
-		div {
-			transform: translateX(0) translateZ(0) !important;
-		}
-
-	}
-
-	.components-editor-notices__dismissible {
-		position: absolute;
-		z-index: 35;
-	}
-}
-
-.components-popover.more-menu-dropdown__content {
-	z-index: z-index(".components-popover.more-menu__content");
 }
