+++ packages/edit-post/src/style.scss
@@ -39,7 +39,7 @@ body.js.block-editor-page {
 }
 
 // Target the editor UI excluding the visual editor contents, metaboxes and custom fields areas.
-.edit-post-header,
+.editor-header,
 .edit-post-text-editor,
 .editor-sidebar,
 .editor-post-publish-panel {
