+++ packages/edit-site/src/components/header-edit-mode/style.scss
@@ -1,192 +1,3 @@
-.edit-site-header-edit-mode {
-	height: $header-height;
-	align-items: center;
-	background-color: $white;
-	color: $gray-900;
-	display: flex;
-	box-sizing: border-box;
-	width: 100%;
-	justify-content: space-between;
-	border-bottom: $border-width solid $gray-200;
+.editor-header {
 	padding-left: $header-height;
-
-	// When top toolbar is engaged and should expand fully.
-	&.show-block-toolbar {
-
-		.edit-site-header-edit-mode__start,
-		.edit-site-header-edit-mode__end {
-			flex-basis: auto;
-		}
-
-		.edit-site-header-edit-mode__center {
-			display: none;
-		}
-	}
-
-	.edit-site-header-edit-mode__start {
-		display: flex;
-		border: none;
-		align-items: center;
-		flex-grow: 1;
-		flex-shrink: 2;
-		// Take up the full height of the header so the border focus
-		// is visible on toolbar buttons.
-		height: 100%;
-		// Allow focus ring to be fully visible on furthest right button.
-		@include break-medium() {
-			padding-right: var(--wp-admin-border-width-focus);
-			// Account for the site hub, which is 60x60px.
-			flex-basis: calc(37.5% - 60px);
-			// We need this to be overflow hidden so the block toolbar can
-			// overflow scroll. If the overflow is visible, flexbox allows
-			// the toolbar to grow outside of the allowed container space.
-			overflow: hidden;
-		}
-	}
-
-	.edit-site-header-edit-mode__end {
-		display: flex;
-		justify-content: flex-end;
-		height: 100%;
-		flex-grow: 1;
-		flex-shrink: 1;
-
-		@include break-medium() {
-			flex-basis: 37.5%;
-		}
-	}
-
-	.edit-site-header-edit-mode__center {
-		align-items: center;
-		display: flex;
-		flex-basis: 100%;
-		flex-grow: 1;
-		flex-shrink: 2;
-		height: 100%;
-		justify-content: center;
-
-		// Flex items will, by default, refuse to shrink below a minimum
-		// intrinsic width. In order to shrink this flexbox item, and
-		// subsequently truncate child text, we set an explicit min-width.
-		// See https://dev.w3.org/csswg/css-flexbox/#min-size-auto
-		min-width: 0;
-
-		@include break-medium() {
-			flex-basis: 25%;
-		}
-	}
-
-}
-
-.edit-site-header-edit-mode__toolbar {
-	align-items: center;
-	display: flex;
-	gap: $grid-unit-10;
-	padding-left: $grid-unit-20;
-
-	@include break-medium() {
-		padding-left: $grid-unit-50 * 0.5;
-	}
-
-	@include break-wide() {
-		padding-right: $grid-unit-10;
-	}
-
-	.edit-site-header-edit-mode__inserter-toggle {
-		svg {
-			transition: transform cubic-bezier(0.165, 0.84, 0.44, 1) 0.2s;
-			@include reduce-motion("transition");
-		}
-
-		&.is-pressed {
-			svg {
-				transform: rotate(45deg);
-			}
-		}
-	}
-}
-
-/**
- * Buttons on the right side
- */
-
-.edit-site-header-edit-mode__actions {
-	display: inline-flex;
-	align-items: center;
-	flex-wrap: nowrap;
-	// Ensure actions do not press against .edit-site-header-edit-mode__center.
-	padding-left: $grid-unit-10;
-	padding-right: $grid-unit-10;
-	gap: $grid-unit-10;
-}
-
-// Button text label styles
-
-.edit-site-header-edit-mode.show-icon-labels {
-	.components-button.has-icon {
-		width: auto;
-
-		// Hide the button icons when labels are set to display...
-		svg {
-			display: none;
-		}
-		// ... and display labels.
-		&::after {
-			content: attr(aria-label);
-		}
-		&[aria-disabled="true"] {
-			background-color: transparent;
-		}
-	}
-	.is-tertiary {
-		&:active {
-			box-shadow: 0 0 0 1.5px var(--wp-admin-theme-color);
-			background-color: transparent;
-		}
-	}
-	// Some margins and padding have to be adjusted so the buttons can still fit on smaller screens.
-	.edit-site-save-button__button {
-		padding-left: 6px;
-		padding-right: 6px;
-	}
-
-	// The template details toggle has a custom label, different from its aria-label, so we don't want to display both.
-	.edit-site-document-actions__get-info.edit-site-document-actions__get-info.edit-site-document-actions__get-info {
-		&::after {
-			content: none;
-		}
-	}
-
-	.edit-site-header-edit-mode__inserter-toggle.edit-site-header-edit-mode__inserter-toggle,
-	.edit-site-document-actions__get-info.edit-site-document-actions__get-info.edit-site-document-actions__get-info {
-		height: 36px;
-		padding: 0 $grid-unit-10;
-	}
-
-	.block-editor-block-mover {
-		// Modified group borders.
-		border-left: none;
-
-		&::before {
-			content: "";
-			width: $border-width;
-			height: $grid-unit-30;
-			background-color: $gray-300;
-			margin-top: $grid-unit-05;
-			margin-left: $grid-unit;
-		}
-
-		// Modified block movers horizontal separator.
-		.block-editor-block-mover__move-button-container {
-			&::before {
-				width: calc(100% - #{$grid-unit-30});
-				background: $gray-300;
-				left: calc(50% + 1px);
-			}
-		}
-	}
-}
-
-.components-popover.more-menu-dropdown__content {
-	z-index: z-index(".components-popover.more-menu__content");
 }
